This PR introduces changes to,
- Run opa filter benchmarks with and without decision logs
- Run the benchmarks in parallel to possibly detect contention issues and overall performance under load
- Measure percentiles in benchmark tests (inspired by https://github.com/zalando/skipper/pull/2699)