@mefarazath 2024-10-21
Sample benchmark run measuring percentiles

```
go test -benchmem -run=^$ -bench ^BenchmarkAuthorizeRequest$ -benchtime=3s
```

```
BenchmarkAuthorizeRequest/minimal/no-decision-logs-10         	  653377	      4862 ns/op	   26142 B/op	     487 allocs/op
BenchmarkAuthorizeRequest/minimal/no-decision-logs/P50-10     	  653377	     31459 ns/op
BenchmarkAuthorizeRequest/minimal/no-decision-logs/P99.5-10   	  653377	    396833 ns/op
BenchmarkAuthorizeRequest/minimal/no-decision-logs/P99.9-10   	  653377	    710041 ns/op
BenchmarkAuthorizeRequest/minimal/with-decision-logs-10       	   62737	     51580 ns/op	   79947 B/op	    1331 allocs/op
BenchmarkAuthorizeRequest/minimal/with-decision-logs/P50-10   	   62737	     97250 ns/op
BenchmarkAuthorizeRequest/minimal/with-decision-logs/P99.5-10 	   62737	  15023917 ns/op
BenchmarkAuthorizeRequest/minimal/with-decision-logs/P99.9-10 	   62737	  15544875 ns/op
BenchmarkAuthorizeRequest/with-body/no-decision-logs-10       	  335612	     78599 ns/op	   35122 B/op	     513 allocs/op
BenchmarkAuthorizeRequest/with-body/no-decision-logs/P50-10   	  335612	    590375 ns/op
BenchmarkAuthorizeRequest/with-body/no-decision-logs/P99.5-10 	  335612	   3332458 ns/op
BenchmarkAuthorizeRequest/with-body/no-decision-logs/P99.9-10 	  335612	   4318250 ns/op
BenchmarkAuthorizeRequest/with-body/with-decision-logs-10     	   61664	     63934 ns/op	   91055 B/op	    1404 allocs/op
BenchmarkAuthorizeRequest/with-body/with-decision-logs/P50-10 	   61664	    240916 ns/op
BenchmarkAuthorizeRequest/with-body/with-decision-logs/P99.5-10	   61664	  14367416 ns/op
BenchmarkAuthorizeRequest/with-body/with-decision-logs/P99.9-10	   61664	  15672583 ns/op
BenchmarkAuthorizeRequest/jwt-validation/no-decision-logs-10  	  211809	     16336 ns/op	   60280 B/op	     964 allocs/op
BenchmarkAuthorizeRequest/jwt-validation/no-decision-logs/P50-10	  211809	    103541 ns/op
BenchmarkAuthorizeRequest/jwt-validation/no-decision-logs/P99.5-10	  211809	   1552625 ns/op
BenchmarkAuthorizeRequest/jwt-validation/no-decision-logs/P99.9-10	  211809	   2376417 ns/op
BenchmarkAuthorizeRequest/jwt-validation/with-decision-logs-10         	   47227	     74450 ns/op	  122129 B/op	    1808 allocs/op
BenchmarkAuthorizeRequest/jwt-validation/with-decision-logs/P50-10     	   47227	    217292 ns/op
BenchmarkAuthorizeRequest/jwt-validation/with-decision-logs/P99.5-10   	   47227	  16026917 ns/op
BenchmarkAuthorizeRequest/jwt-validation/with-decision-logs/P99.9-10   	   47227	  17956916 ns/op
```