@AlexanderYastrebov 2024-10-21
> best would be to show benchstat results like #2876

See https://pkg.go.dev/golang.org/x/perf/cmd/benchstat