@mefarazath 2024-10-21
@szuecs @AlexanderYastrebov

Apologies for the confusion. This PR simply improves the existing benchmark code to add more measurements to get better insights into latency distribution and does not add anything to improve the performance. Added sub benchmarks also helps us to get an idea for each scenario with and without decision logging enabled.

This would gives us a quick way to verify/validate performance improvements proposed to filter and opa with better clarity.

As an example I did a comparison between v0.68.0 (current version in skipper) vs a [PoC](https://github.com/open-policy-agent/opa/issues/5724#issuecomment-2424067534) done by OPA team to improve decision logs performance. The p995 and p999 gives a better idea on the impact of the improvements here.

```
➜  benchmark-results benchstat v0.68.0.txt after_fix.txt                                                                                                                              7ms
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/filters/openpolicyagent/opaauthorizerequest
cpu: Apple M1 Max
                                                            │  v0.68.0.txt   │            after_fix.txt             │
                                                            │     sec/op     │    sec/op     vs base                │
AuthorizeRequest/minimal/no-decision-logs-10                    4.798µ ±  8%   4.831µ ±  1%   +0.68% (p=0.043 n=10)
AuthorizeRequest/minimal/no-decision-logs/P50-10                31.23µ ±  0%   31.46µ ±  0%   +0.73% (p=0.003 n=10)
AuthorizeRequest/minimal/no-decision-logs/P99-10                303.9µ ± 14%   299.6µ ±  4%        ~ (p=0.424 n=10)
AuthorizeRequest/minimal/no-decision-logs/P99.5-10              389.0µ ± 31%   382.2µ ±  5%        ~ (p=0.218 n=10)
AuthorizeRequest/minimal/no-decision-logs/P99.9-10              694.5µ ± 53%   682.1µ ±  4%        ~ (p=0.247 n=10)

AuthorizeRequest/minimal/with-decision-logs-10                  51.18µ ±  1%   11.67µ ±  1%  -77.21% (p=0.000 n=10)
AuthorizeRequest/minimal/with-decision-logs/P50-10              96.92µ ±  2%   98.71µ ±  1%   +1.85% (p=0.034 n=10)
AuthorizeRequest/minimal/with-decision-logs/P99-10            14679.2µ ±  1%   417.6µ ±  1%  -97.16% (p=0.000 n=10)
AuthorizeRequest/minimal/with-decision-logs/P99.5-10          14983.9µ ±  1%   504.6µ ±  2%  -96.63% (p=0.000 n=10)
AuthorizeRequest/minimal/with-decision-logs/P99.9-10          15417.1µ ±  2%   762.6µ ±  5%  -95.05% (p=0.000 n=10)

AuthorizeRequest/with-body/no-decision-logs-10                  68.94µ ±  6%   82.72µ ± 13%  +19.99% (p=0.001 n=10)
AuthorizeRequest/with-body/no-decision-logs/P50-10              516.0µ ±  5%   622.4µ ± 12%  +20.62% (p=0.001 n=10)
AuthorizeRequest/with-body/no-decision-logs/P99-10              2.468m ±  5%   2.851m ±  9%  +15.52% (p=0.000 n=10)
AuthorizeRequest/with-body/no-decision-logs/P99.5-10            2.762m ±  5%   3.237m ± 10%  +17.22% (p=0.000 n=10)
AuthorizeRequest/with-body/no-decision-logs/P99.9-10            3.477m ±  6%   4.131m ± 21%  +18.80% (p=0.000 n=10)

AuthorizeRequest/with-body/with-decision-logs-10                62.09µ ±  1%   55.19µ ± 10%  -11.11% (p=0.002 n=10)
AuthorizeRequest/with-body/with-decision-logs/P50-10            244.9µ ±  1%   424.0µ ±  7%  +73.11% (p=0.000 n=10)
AuthorizeRequest/with-body/with-decision-logs/P99-10           13.792m ±  1%   1.809m ± 13%  -86.88% (p=0.000 n=10)
AuthorizeRequest/with-body/with-decision-logs/P99.5-10         14.158m ±  1%   2.065m ± 11%  -85.42% (p=0.000 n=10)
AuthorizeRequest/with-body/with-decision-logs/P99.9-10         15.205m ±  2%   2.859m ± 23%  -81.20% (p=0.000 n=10)

AuthorizeRequest/jwt-validation/no-decision-logs-10             15.62µ ±  2%   14.45µ ±  5%   -7.50% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/no-decision-logs/P50-10         102.1µ ±  1%   100.7µ ±  0%   -1.37% (p=0.001 n=10)
AuthorizeRequest/jwt-validation/no-decision-logs/P99-10        1023.9µ ±  3%   779.0µ ± 30%  -23.91% (p=0.004 n=10)
AuthorizeRequest/jwt-validation/no-decision-logs/P99.5-10       1.493m ±  2%   1.263m ± 29%        ~ (p=0.123 n=10)
AuthorizeRequest/jwt-validation/no-decision-logs/P99.9-10       2.226m ±  6%   3.536m ± 18%  +58.80% (p=0.000 n=10)

AuthorizeRequest/jwt-validation/with-decision-logs-10           70.05µ ±  1%   22.51µ ±  2%  -67.87% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/with-decision-logs/P50-10       201.6µ ±  3%   192.5µ ±  2%   -4.52% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/with-decision-logs/P99-10     15299.9µ ±  1%   654.3µ ±  4%  -95.72% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/with-decision-logs/P99.5-10   15798.8µ ±  1%   780.5µ ±  4%  -95.06% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/with-decision-logs/P99.9-10    16.547m ±  1%   1.226m ± 33%  -92.59% (p=0.000 n=10)
geomean                                                         818.0µ         345.9µ        -57.71%

                                                      │  v0.68.0.txt  │            after_fix.txt             │
                                                      │     B/op      │     B/op      vs base                │
AuthorizeRequest/minimal/no-decision-logs-10             25.53Ki ± 0%   25.68Ki ± 0%   +0.61% (p=0.000 n=10)
AuthorizeRequest/minimal/with-decision-logs-10           78.01Ki ± 0%   52.79Ki ± 0%  -32.33% (p=0.000 n=10)
AuthorizeRequest/with-body/no-decision-logs-10           34.30Ki ± 0%   34.46Ki ± 0%   +0.46% (p=0.000 n=10)
AuthorizeRequest/with-body/with-decision-logs-10         88.93Ki ± 0%   62.42Ki ± 0%  -29.81% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/no-decision-logs-10      58.87Ki ± 0%   59.14Ki ± 0%   +0.46% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/with-decision-logs-10   119.34Ki ± 0%   89.59Ki ± 0%  -24.93% (p=0.000 n=10)
geomean                                                  59.12Ki        49.91Ki       -15.58%

                                                      │ v0.68.0.txt │             after_fix.txt             │
                                                      │  allocs/op  │  allocs/op   vs base                  │
AuthorizeRequest/minimal/no-decision-logs-10             487.0 ± 0%    487.0 ± 0%        ~ (p=1.000 n=10) ¹
AuthorizeRequest/minimal/with-decision-logs-10          1.330k ± 0%   1.042k ± 0%  -21.65% (p=0.000 n=10)
AuthorizeRequest/with-body/no-decision-logs-10           513.0 ± 0%    513.0 ± 0%        ~ (p=1.000 n=10) ¹
AuthorizeRequest/with-body/with-decision-logs-10        1.404k ± 0%   1.096k ± 0%  -21.94% (p=0.000 n=10)
AuthorizeRequest/jwt-validation/no-decision-logs-10      964.0 ± 0%    964.0 ± 0%        ~ (p=1.000 n=10) ¹
AuthorizeRequest/jwt-validation/with-decision-logs-10   1.809k ± 0%   1.520k ± 0%  -15.98% (p=0.000 n=10)
geomean                                                  966.2         864.7       -10.50%
¹ all samples are equal
```